+++ MoneyFlow/app/src/main/java/com/derlars/moneyflow/Database/Value.java
@@ -9,66 +9,60 @@ public class Value<V extends Comparable> extends BaseValue {
     private V databaseValue;
     private V value;
 
-    protected boolean readOnly;
-
-    public Value(String path, String key, BaseValueCallback callback) {
-        this(path, key, callback,false);
-    }
-
-    public Value(String path, String key, BaseValueCallback callback, boolean readOnly) {
-        super(path, key, callback);
-        this.readOnly = readOnly;
+    public Value(String path, String key, boolean readable, boolean writable, boolean connectOnRequest, BaseValueCallback callback) {
+        super(path, key, readable, writable,connectOnRequest, callback);
     }
 
     public void set(V v) {
-        if(!readOnly || !online) {
+        if(databaseValue == null || writable) {
+            //Log.d("UNITTEST","Setting value for " + this.getClass().getName() + " -> " + this.path + "/" + this.key + ": " + v);
             value = v;
         }
 
-        if(!readOnly && online && connected && (databaseValue == null || value.compareTo(databaseValue) != 0)) {
+        if(writable && online && connected && (databaseValue == null || value.compareTo(databaseValue) != 0)){
             database.setValue(v);
         }
     }
 
     public V get() {
+        database.getValue();
+
+        //Log.d("UNITTEST","Getting value for " + this.getClass().getName() + " -> " + this.path + "/" + this.key + ": " + value);
         return value;
     }
 
     public void delete() {
         value = null;
 
-        if(online) {
+        if(writable && online) {
             database.deleteValue();
         }
     }
 
-    private void update() {
-        if(online) {
-            value = (V)database.getValue();
-            if(value != null && (databaseValue == null || value.compareTo(databaseValue) != 0)) {
+    private void updateValue() {
+        if(readable && online) {
+            V tmp = (V)database.getValue();
+
+            //Log.d("UNITTEST","Updating value for " + this.getClass().getName() + " -> " + this.path + "/" + this.key + ": " + value);
+
+            if(tmp != null && (databaseValue == null || tmp.compareTo(databaseValue) != 0)) {
+                value = tmp;
                 notifyUpdate(key);
-            }else if(value == null) {
+            }else if(tmp == null) {
                 notifyNotOnline(key);
             }
-        }else{
-            Log.d("MUTEX",this.key + " is not considered as online");
         }
 
         databaseValue = (V)database.getValue();
-    }
-
-    public void setReadOnly(boolean readOnly) {
-        this.readOnly = readOnly;
     }
 
     @Override
     public void setOnline() {
         if(!online) {
             super.setOnline();
-            if(isConnected()) {
+
+            if(writable && isConnected()) {
                 database.setValue(value);
-            }else{
-                Log.d("MUTEX","not connected...");
             }
         }
     }
@@ -78,7 +72,7 @@ public class Value<V extends Comparable> extends BaseValue {
         super.databaseValueRetrieved(path, key, dataSnapshot);
 
         settingOnline = false;
-        update();
+        updateValue();
     }
 
     @Override
@@ -97,11 +91,12 @@ public class Value<V extends Comparable> extends BaseValue {
     public void databaseValueDeleted(String path, String key) {
         super.databaseValueDeleted(path, key);
 
-        update();
+        updateValue();
     }
 
     @Override
     public void update(String key) {
-
+        setOnline();
+        notifyUpdate(this.key);
     }
 }
